migration of data building to store
