+++ src/main/java/codeu/controller/ActivityFeedServlet.java
@@ -25,8 +25,6 @@ public class ActivityFeedServlet extends HttpServlet {
   /** Store class that gives access to Users. */
   private UserStore userStore;
 
-  ArrayList<Instant> eventsInstantsSorted = new ArrayList<Instant>();
-
   /** Set up state for handling chat requests. */
   @Override
   public void init() throws ServletException {
@@ -60,40 +58,8 @@ public class ActivityFeedServlet extends HttpServlet {
     this.userStore = userStore;
   }
 
-  /**
-   * This function takes in a list of conversations,users and messages and returns a HashMap of
-   * Instants mapping to HashMaps of UUIDs mapping to a data store string: User,Conversation or
-   * Message
-   * 
-   * @param conversations
-   * @param users
-   * @param messages
-   * @return built HashMap
-   */
-  HashMap<Instant, HashMap<UUID, String>> buildHashMap(List<Conversation> conversations,
-      List<User> users, List<Message> messages) {
-    HashMap<Instant, HashMap<UUID, String>> outerhm = new HashMap<Instant, HashMap<UUID, String>>();
 
-    for (User user : users) {
-      HashMap<UUID, String> innerhm = new HashMap<UUID, String>();
-      innerhm.put(user.getId(), "user");
-      outerhm.put(user.getCreationTime(), innerhm);
 
-    }
-
-    for (Conversation conversation : conversations) {
-      HashMap<UUID, String> innerhm = new HashMap<UUID, String>();
-      innerhm.put(conversation.getId(), "conversation");
-      outerhm.put(conversation.getCreationTime(), innerhm);
-
-    }
-    for (Message message : messages) {
-      HashMap<UUID, String> innerhm = new HashMap<UUID, String>();
-      innerhm.put(message.getId(), "message");
-      outerhm.put(message.getCreationTime(), innerhm);
-    }
-    return outerhm;
-  }
 
   /**
    * This function takes in a HashMap and finds the earliest Instant using the isBefore function
@@ -101,20 +67,7 @@ public class ActivityFeedServlet extends HttpServlet {
    * @param hm
    * @return the earliest instant
    */
-  Instant findEarliestInstant(HashMap<Instant, HashMap<UUID, String>> hm) {
-    Instant earliestInstant = null;
-    for (Map.Entry<Instant, HashMap<UUID, String>> m : hm.entrySet()) {
-      earliestInstant = m.getKey();
-      break;
-    }
-    for (Map.Entry<Instant, HashMap<UUID, String>> m : hm.entrySet()) {
-      Instant toCmpInstant = m.getKey();
-      if (earliestInstant.isBefore(toCmpInstant)) {
-        earliestInstant = toCmpInstant;
-      }
-    }
-    return earliestInstant;
-  }
+
 
   /**
    * This function takes in a HashMap and sorts the instants from latest to oldest Stores Instants
@@ -123,21 +76,7 @@ public class ActivityFeedServlet extends HttpServlet {
    * @param hm
    * @return a sorted HashMap by Instants
    */
-  HashMap<Instant, HashMap<UUID, String>> sortHashMap(HashMap<Instant, HashMap<UUID, String>> hm) {
-    Instant earlier = null;
-    int size = hm.size();
-    eventsInstantsSorted = new ArrayList<Instant>();
-    HashMap<Instant, HashMap<UUID, String>> sortedhm =
-        new HashMap<Instant, HashMap<UUID, String>>();
-    for (int i = 0; i < size; i++) {
-      earlier = findEarliestInstant(hm);
-      eventsInstantsSorted.add(earlier);
-      HashMap<UUID, String> innersm = hm.get(earlier);
-      sortedhm.put(earlier, innersm);
-      hm.remove(earlier);
-    }
-    return sortedhm;
-  }
+
 
   /**
    * This function fires when a user requests the /activity URL. It simply forwards the request to
@@ -149,19 +88,25 @@ public class ActivityFeedServlet extends HttpServlet {
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response)
       throws ServletException, IOException {
+    ArrayList<Instant> eventsInstantsSorted = new ArrayList<Instant>();
+    HashMap<Instant, HashMap<UUID, String>> eventsMap =
+        new HashMap<Instant, HashMap<UUID, String>>();
+
     List<User> users = userStore.getAllUsers();
     List<Conversation> conversations = conversationStore.getAllConversations();
     List<Message> messages = messageStore.getAllMessages();
 
-    HashMap<Instant, HashMap<UUID, String>> eventsMap =
-        buildHashMap(conversations, users, messages);
-    eventsMap = sortHashMap(eventsMap);
-    userStore.setEventsInstants(eventsInstantsSorted);
+    eventsMap = userStore.buildEventsMap(users, conversations, messages);
+    eventsMap = userStore.sortEventsMap(eventsMap);
+    eventsInstantsSorted = userStore.getAllEventsInstants();
+
     request.setAttribute("users", users);
     request.setAttribute("conversations", conversations);
     request.setAttribute("messages", messages);
-    // request.setAttribute("eventsInstantsSorted", eventsInstantsSorted);
-    // request.setAttribute("eventsMap", eventsMap);
+
+    request.setAttribute("eventsInstantsSorted", eventsInstantsSorted);
+    request.setAttribute("eventsMap", eventsMap);
+
     request.getRequestDispatcher("/WEB-INF/view/activityfeed.jsp").forward(request, response);
   }
